dom/url/tests/test_worker_url_exceptions.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE HTML>
<html>
<head>
  <title>Test for URL exceptions in workers</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none"></div>
<pre id="test"></pre>
<script class="testbody" type="text/javascript">

  var worker = new Worker("url_exceptions_worker.js");

  worker.onmessage = function(event) {
    is(event.target, worker);

    if (event.data.type == "finish") {
      SimpleTest.finish();
    } else if (event.data.type == "status") {
      ok(event.data.status, event.data.msg);
    }
  };

  worker.onerror = function(event) {
    is(event.target, worker);
    ok(false, "Worker had an error: " + event.message);
    SimpleTest.finish();
  };

  worker.postMessage(0);
  SimpleTest.waitForExplicitFinish();

</script>
</pre>
</body>
</html>
